LayoutTests/editing/spelling/spellcheck-paste-disabled.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
<script src="resources/js-test-selection-shared.js"></script>
</head>
<body>
<div>
<div id="src" contenteditable="true" spellcheck="true"></div><br/>
<div id="dst" contenteditable="true" spellcheck="false"></div>
</div>
<script language="javascript">
description('Test if WebKit does not spellcheck text when pasting text to an element having spellchecking disabled. ' +
            'To test manually, copy the text "zz apple" and paste it to the textarea. ' +
            'When "zz" is not marked as misspelled, this test succeeds.');

// Insert a misspelled word to initialize a spellchecker.
var srcNode = document.getElementById('src');
srcNode.focus();
document.execCommand("InsertText", false, 'z');
document.execCommand("InsertText", false, 'z');
document.execCommand("InsertText", false, ' ');
document.execCommand("InsertText", false, 'a');
document.execCommand("InsertText", false, 'p');
document.execCommand("InsertText", false, 'p');
document.execCommand("InsertText", false, 'l');
document.execCommand("InsertText", false, 'e');
shouldBeTrue('internals.hasSpellingMarker(0, 2)');
window.getSelection().selectAllChildren(srcNode);
document.execCommand('Copy');

var dstNode = document.getElementById('dst');
dstNode.focus();
document.execCommand('Paste');
shouldBeFalse('internals.hasSpellingMarker(0, 2)');

</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
